+++ src/main/java/io/hhplus/tdd/point/PointHistory.java
@@ -1,5 +1,7 @@
 package io.hhplus.tdd.point;
 
+import lombok.Builder;
+
 public record PointHistory(
         long id,
         long userId,
